requirements.txt
Django==3.0.8
djangorestframework==3.11.1
